requirements.txt
streamlit>=1.30.0
python-dotenv>=1.0.0
rapidfuzz>=3.5.0
nltk>=3.8.1
num2words>=0.5.13
pyyaml>=6.0.1
pydub>=0.25.1
assemblyai>=0.17.0
openai>=1.0.0
pydantic>=2.0.0

